requirements.txt
geodezyx @ git+https://github.com/IPGP/geodezyx
rinexmod @ git+https://github.com/IPGP/rinexmod
dateparser
docker
filelock
numpy
pandas
PyYAML
python-dateutil
termcolor
hatanaka
requests
tqdm
bs4
beautifulsoup4
colorlog
tabulate
matplotlib
jinja2
colorama
logging-tree
mergedeep